split similarity table helpers
